S3 object functions for idml class objects

new_idml() creates a idml class object from a file path and contents. validate_idml() checks idml objects. Used by read_idml() to create idml objects from a file and by other functions to validate input idml objects.

Usage

new_idml(file, path, contents, error_call = caller_env())

validate_idml(
  idml,
  what = "idml",
  nm = c("file", "path", "contents"),
  type = "application/vnd.adobe.indesign-idml-package",
  arg = caller_arg(idml),
  error_call = caller_env()
)

Arguments

file

File name for input IDML file.

path

Path to location of unzipped IDML file. This is the directory where the XML files are located.

contents

A named list of xml_document objects.

error_call

The execution environment of a currently running function, e.g. caller_env(). The function will be mentioned in error messages as the source of the error. See the call argument of abort() for more information.

idml

An idml class object.

type

MIMETYPE value to use in validating idml objects.

arg

An argument name as a string. This argument will be mentioned in error messages as the input that is at the origin of a problem.