List stories from the content of an idml object

idml_list_stories() lists stories and related elements from a idml object. idml_list_story_paras() returns just the ParagraphStyleRange element of the stories.

Usage

idml_list_stories(idml, format = "contents")

idml_list_story_paras(idml, ...)

Arguments

idml

An idml class object.

format

Format of output to return. Must be "contents"

Value

A bare named list with four elements: Stories, StoryPreference, InCopyExportOption, and ParagraphStyleRange.

Examples

if (FALSE) { # \dontrun{
if(interactive()){
  path <- system.file("idml/letter_portrait_standard.idml", package = "idmlr")

  idml_obj <- read_idml(path)

  idml_list_stories(idml_obj)
 }
} # }