List styles from an idml object

idml_list_styles() takes an idml object and returns a bare list with names, properties, attributes, and content from the XML nodes definition the paragraph, character, or other styles of the specified type.

Usage

idml_list_styles(..., type = "paragraph", format = "list")

Arguments

...

Arguments passed on to get_idml_styles

idml

A idml class object created with read_idml().

parent_nm

Name of top-level parent node. If parent_nm is a character vector, extract_doc_df() returns a data frame list.

type

Type of data to extract from chlildren nodes: "attr" (attributes) or "text".

format

Output format must be "list"

Value

A bare list with four elements: a named list of style attributes (names are style names), properties for the style XML nodes, attributes for the style XML nodes, and the contents of each XML node.

Examples

if (FALSE) { # \dontrun{
if(interactive()){
  path <- system.file("idml/letter_portrait_standard.idml", package = "idmlr")

  idml_obj <- read_idml(path)

  idml_list_styles(idml_obj)
 }
} # }