Read an InDesign Markup Language (IDML) file to a idml object

read_idml() takes a file path for an idml file and returns an idml object: a list containing a file name, file path, and the contents of the file as a list of xml_document objects and character vectors (for the MIMETYPE file only).

Usage

read_idml(file, ..., exdir = tempdir())

Arguments

file

Path to a file ending in a .idml file extension.

...

Additional parameters passed to xml2::read_xml()

exdir

Directory to extract files to when using utils::unzip(), Default: tempdir().

Value

A idml class object.

See also