Create a data frame from an XML document

Based on https://rud.is/rpubs/xml2power/

Usage

xml_doc_to_df(
  doc = NULL,
  parent_nm = NULL,
  parent_node = NULL,
  node_names = NULL,
  unique_nm = TRUE,
  nested = TRUE,
  type = "attr",
  ...
)

Arguments

doc

A XML document.

parent_nm

Name of top-level parent node. If parent_nm is a character vector, [extract_doc_df()] returns a data frame list.

parent_node

A XML document, node, or node set.

node_names

Node names to iterate over. Default `NULL`. Typically created from children of parent node.

unique_nm

If `TRUE`, ensure that parent_node names are unique before using them to extract data from the `xml_document`. Ignored if `node_names` is supplied.

nested

If `TRUE`, use `parent_nm` to extract a parent node from doc. If `FALSE`, `parent_node` defaults to doc (unless `parent_node` is supplied).

type

Type of data to extract from chlildren nodes: "attr" (attributes) or "text".

...

Arguments passed on to readr::type_convert

df

A data frame.

col_types

One of NULL, a cols() specification, or a string. See vignette("readr") for more details.

If NULL, column types will be imputed using all rows.

na

Character vector of strings to interpret as missing values. Set this option to character() to indicate no missing values.

trim_ws

Should leading and trailing whitespace (ASCII spaces and tabs) be trimmed from each field before parsing it?

locale

The locale controls defaults that vary from place to place. The default locale is US-centric (like R), but you can use locale() to create your own locale that controls things like the default time zone, encoding, decimal mark, big mark, and day/month names.

guess_integer

If TRUE, guess integer types for whole numbers, if FALSE guess numeric type for all numbers.